convert_to_notebook.py
import nbformat as nbf

# Create a new notebook
nb = nbf.v4.new_notebook()

# Read the Python file
with open('deepfake_detection.py', 'r') as f:
    code = f.read()

# Split the code into sections
sections = code.split('\n\n')

# Add title and setup instructions
title_cell = nbf.v4.new_markdown_cell('''# Deepfake Detection Project

This notebook implements a deepfake detection system using a pre-trained model. The system can detect manipulated images and videos in real-time.

## Setup Instructions

1. Install required dependencies:
```bash
pip install -r requirements.txt
```

2. Download the pre-trained model (will be done automatically)

3. Run all cells in sequence

## Project Structure

- `templates/`: Contains the HTML template
- `static/`: Contains CSS and JavaScript files
- `project.ipynb`: This notebook with the implementation''')
nb.cells.append(title_cell)

# Add imports section
imports_cell = nbf.v4.new_code_cell(sections[0])
nb.cells.append(imports_cell)

# Add model loading section
model_title = nbf.v4.new_markdown_cell('''## Load Pre-trained Model

We'll use a pre-trained EfficientNet model for deepfake detection. The model has been trained on a large dataset of real and manipulated images.''')
nb.cells.append(model_title)

model_cell = nbf.v4.new_code_cell(sections[1])
nb.cells.append(model_cell)

# Add image processing section
processing_title = nbf.v4.new_markdown_cell('''## Image Processing Functions

These functions handle image preprocessing and feature extraction.''')
nb.cells.append(processing_title)

processing_cell = nbf.v4.new_code_cell(sections[2])
nb.cells.append(processing_cell)

# Add detection section
detection_title = nbf.v4.new_markdown_cell('''## Detection Functions

These functions handle the actual deepfake detection process.''')
nb.cells.append(detection_title)

detection_cell = nbf.v4.new_code_cell(sections[3])
nb.cells.append(detection_cell)

# Add Flask routes section
routes_title = nbf.v4.new_markdown_cell('''## Flask Routes

These routes handle the web interface and API endpoints.''')
nb.cells.append(routes_title)

routes_cell = nbf.v4.new_code_cell(sections[4])
nb.cells.append(routes_cell)

# Add run section
run_title = nbf.v4.new_markdown_cell('''## Run the Application

Start the Flask server to run the application.''')
nb.cells.append(run_title)

run_cell = nbf.v4.new_code_cell(sections[5])
nb.cells.append(run_cell)

# Add testing section
testing_title = nbf.v4.new_markdown_cell('''## Testing the Application

You can now test the application by:

1. Running all cells in this notebook
2. Opening a web browser and navigating to `http://localhost:5000`
3. Uploading an image or video file through the interface
4. Clicking the 'Analyze' button to get the detection results

The application will provide:
- A binary classification (Deepfake/Not Deepfake)
- A confidence score for the prediction
- Visual feedback through the web interface''')
nb.cells.append(testing_title)

# Write the notebook
with open('project.ipynb', 'w') as f:
    nbf.write(nb, f) 