requirements.txt
flask==2.3.3
tensorflow-cpu==2.10.0
opencv-python==4.8.0.74
numpy==1.24.3
matplotlib==3.7.2
werkzeug==2.3.7 